+++ changelog.md
@@ -1,3 +1,8 @@
+### v1.1.5
+[x] Custom Keybinds support for anything but mac because it doesn't work yet apparently
+[x] Animated Background toggler added!
+[ ] Setting to change to Undertale background added!
+
 ### v1.1.4
 - A little bit of cleanup
 - Crashfix